requirements.txt
absl-py==0.11.0
adversarial-robustness-toolbox==1.2.0
anykeystore==0.2
apex==0.9.10.dev0
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2020.6.20
chardet==3.0.4
cryptacular==1.5.5
cycler==0.10.0
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
grpcio==1.34.0
hupper==1.10.2
idna==2.8
importlib-metadata==3.1.1
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.1
PasteDeploy==2.1.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plaster==1.0
plaster-pastedeploy==0.7
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.14.0
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
pyramid==1.10.5
pyramid-mailer==0.15.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.0
QtPy==1.9.0
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.20
tensorboard==1.15.0
terminado==0.8.3
testpath==0.4.4
torch==1.7.1
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.8.2
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
transaction==3.0.1
translationstring==1.4
typing-extensions==3.7.4.3
urllib3==1.25.8
velruse==1.1.1
venusian==3.0.0
visdom==0.1.8.9
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
zipp==3.4.0
zope.deprecation==4.4.0
zope.interface==5.2.0
zope.sqlalchemy==1.3
